Add test for onvif url and move camera tests in modules
